+++ CHANGELOG.md
@@ -12,6 +12,7 @@ glance - what to expact from upgrading to a new version.
 - Importer: Store map names.
 - Importer: Store marine and alien starting locations.
 - Importer: Store time spent per player for each class (lifeform, weapon, ...)
+- Integration of sentry.io for excecption tracking.
 
 ### Changed